version bump

+++ Forge/src/main/java/software/bernie/geckolib3/util/GeckoLibUtil.java
@@ -84,7 +84,7 @@ public class GeckoLibUtil {
 	}
 
 	public static AnimationController getControllerForID(AnimationFactory factory, Integer id, String controllerName) {
-		return factory.getOrCreateAnimationData(id).getAnimationControllers().get(controllerName);
+		return factory.getOrCreateAnimationData(id.intValue()).getAnimationControllers().get(controllerName);
 	}
 
 	/**
